pyproject.toml
[project]
name = "aisuite"
version = "0.1.11"
description = "Uniform access layer for LLMs"
authors = [{ name = "Andrew Ng, Rohit P" }]
requires-python = "~=3.10"
readme = "README.md"
dependencies = [
    "docstring-parser>=0.14.0,<0.15",
    "httpx~=0.27.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.30.1,<0.31"]
aws = ["boto3>=1.34.144,<2"]
azure = []
cerebras = ["cerebras_cloud_sdk>=1.19.0,<2"]
cohere = ["cohere>=5.12.0,<6"]
deepseek = ["openai>=1.35.8,<2"]
google = ["vertexai>=1.63.0,<2"]
googlegenai = ["google-generativeai>=0.8.0,<0.9"]
groq = ["groq>=0.9.0,<0.10"]
huggingface = []
mistral = ["mistralai>=1.0.3,<2"]
ollama = []
openai = ["openai>=1.35.8,<2"]
watsonx = ["ibm-watsonx-ai>=1.1.16,<2"]
all = [
    "anthropic>=0.30.1,<0.31",
    "cerebras_cloud_sdk>=1.19.0,<2",
    "groq>=0.9.0,<0.10",
    "openai>=1.35.8,<2",
    "cohere>=5.12.0,<6",
    "google-generativeai>=0.8.0,<0.9"
]

[dependency-groups]
dev = [
    "pre-commit>=3.7.1,<4",
    "black>=24.4.2,<25",
    "python-dotenv>=1.0.1,<2",
    "openai>=1.35.8,<2",
    "groq>=0.9.0,<0.10",
    "anthropic>=0.30.1,<0.31",
    "notebook>=7.2.1,<8",
    "ollama>=0.2.1,<0.3",
    "mistralai>=1.0.3,<2",
    "boto3>=1.34.144,<2",
    "fireworks-ai>=0.14.0,<0.15",
    "chromadb>=0.5.4,<0.6",
    "sentence-transformers>=3.0.1,<4",
    "datasets>=2.20.0,<3",
    "vertexai>=1.63.0,<2",
    "ibm-watsonx-ai>=1.1.16,<2",
    "cerebras_cloud_sdk>=1.19.0,<2",
]
test = [
    "pytest>=8.2.2,<9",
    "pytest-cov>=6.0.0,<7",
    "respx"
]

[tool.uv]
default-groups = [
    "dev",
    "test",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths="tests"
markers = [
    "integration: marks tests as integration tests that interact with external services",
]
ignore = [
    "tests/client/test_prerelease.py"
]
